refactor: move addCommonButtons to DeviceInfo
