+++ src/types/devices/router.ts
@@ -9,18 +9,6 @@ export class Router extends Device {
     super(id, RouterImage, viewgraph, position);
   }
 
-  showInfo() {
-    const info = new DeviceInfo("Router");
-    info.addField("ID", this.id.toString());
-    info.addListField(
-      "Connected Devices",
-      Array.from(this.connections.values()),
-    );
-    RightBar.getInstance().renderInfo(info);
-
-    this.addCommonButtons();
-  }
-
   getLayer(): Layer {
     return Layer.Network;
   }
