+++ src/types/devices/server.ts
@@ -9,18 +9,6 @@ export class Server extends Device {
     super(id, ServerImage, viewgraph, position);
   }
 
-  showInfo() {
-    const info = new DeviceInfo("Server");
-    info.addField("ID", this.id.toString());
-    info.addListField(
-      "Connected Devices",
-      Array.from(this.connections.values()),
-    );
-    RightBar.getInstance().renderInfo(info);
-
-    this.addCommonButtons();
-  }
-
   getLayer(): Layer {
     return Layer.App;
   }
